Updated the Page to be much more production quality.
Site isn't a complete eyesore to look at and functions properly with added features like cookie management, usernames, and a username changer, as well as a dynamic welcoming system that welcomes the user differently whether they have visited the site or they are new.

And a lil' easter egg. TeeHee